refactor: moved dialogs to components folder.
