+++ src/components/GenerateAccountDialog.vue
