+++ src/components/SetKeystorePasswordDialog.vue
